CLOUD RUN/requirements.txt
pandas
google-cloud-bigquery
google-cloud-bigquery-storage
google-cloud-storage
scikit-learn
joblib
db-dtypes